Move desmume wrapper to its own module
